support for different data types - publisher and subscriber (still testing)

+++ MRML/vtkMRMLROS2PublisherDefaultNodes.cxx
@@ -6,12 +6,11 @@
 VTK_MRML_ROS_PUBLISHER_NATIVE_CXX(std::string, std_msgs::msg::String, String);
 VTK_MRML_ROS_PUBLISHER_NATIVE_CXX(bool, std_msgs::msg::Bool, Bool);
 VTK_MRML_ROS_PUBLISHER_NATIVE_CXX(int, std_msgs::msg::Int64, Int);
-VTK_MRML_ROS_PUBLISHER_NATIVE_CXX(float, std_msgs::msg::Float64, Float);
+VTK_MRML_ROS_PUBLISHER_NATIVE_CXX(double, std_msgs::msg::Float64, Float);
 
 VTK_MRML_ROS_PUBLISHER_VTK_CXX(vtkIntArray, std_msgs::msg::Int64MultiArray, IntArray);
 VTK_MRML_ROS_PUBLISHER_VTK_CXX(vtkDoubleArray, std_msgs::msg::Float64MultiArray, FloatArray);
 VTK_MRML_ROS_PUBLISHER_VTK_CXX(vtkDenseArray<int>, std_msgs::msg::Int64MultiArray, IntNArray);
-
 
 VTK_MRML_ROS_PUBLISHER_VTK_CXX(vtkMatrix4x4, geometry_msgs::msg::PoseStamped, PoseStamped);
 VTK_MRML_ROS_PUBLISHER_VTK_CXX(vtkDoubleArray, geometry_msgs::msg::WrenchStamped, WrenchStamped);
